chore: cargo clippy --fix + fmt sweep (mechanical fixes only)
Applies the machine-applicable suggestions across the workspace (mostly
mismatched_lifetime_syntaxes elision). Takes the -D warnings error count
from ~94 to 29; the remainder are structural calls left for a dedicated
pass: not_unsafe_ptr_arg_deref on nine public fns (an API-signature
change), arc_with_non_send_sync, type_complexity, enum variant sizes.
No behavior change; unit + gpu phases green after the sweep.

+++ wayland-display-core/src/utils/vulkan_share.rs
@@ -198,8 +198,8 @@ unsafe fn physical_index_for_minor(
             let mut drm = vk::PhysicalDeviceDrmPropertiesEXT::default();
             let mut p2 = vk::PhysicalDeviceProperties2::default().push_next(&mut drm);
             ash_inst.get_physical_device_properties2(d, &mut p2);
-            if (drm.has_render != 0 && drm.render_minor as i64 == minor as i64)
-                || (drm.has_primary != 0 && drm.primary_minor as i64 == minor as i64)
+            if (drm.has_render != 0 && drm.render_minor == minor as i64)
+                || (drm.has_primary != 0 && drm.primary_minor == minor as i64)
             {
                 tracing::info!(
                     "vulkan_share: selected physical device {i} '{}' (matches render minor {minor}, video-encode capable)",
